Move flask setup and mf schema in orion folder
